+++ tests/Classes/Database/Query/FromTest.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Sys25\RnBase\Database;
+namespace Sys25\RnBase\Database\Query;
 
 /***************************************************************
 *  Copyright notice
 *
-*  (c) 2010 Rene Nitzsche (rene@system25.de)
+*  (c) 2010-2021 Rene Nitzsche (rene@system25.de)
 *  All rights reserved
 *
 *  This script is part of the TYPO3 project. The TYPO3 project is
